Introduce (or rather actually use) the parse error
