+++ examples/ast.why
@@ -1,0 +1,1 @@
+let a := 2 {}}
