+++ src/ast/parser.rs
@@ -1,18 +1,14 @@
-use log::error;
 use pest::{iterators::Pairs, Parser};
+
+use super::parse_error::{ParseError, ParseResult};
 
 #[derive(Parser)]
 #[grammar = "y-lang.pest"]
 pub struct YParser;
 
 impl YParser {
-    pub fn parse_program(program: &str) -> Pairs<Rule> {
-        match Self::parse(Rule::program, program) {
-            Ok(pairs) => pairs,
-            Err(err) => {
-                error!("Failed to parse file ({})", err);
-                std::process::exit(-1);
-            }
-        }
+    pub fn parse_program(file: impl ToString, program: &str) -> ParseResult<Pairs<Rule>> {
+        Self::parse(Rule::program, program)
+            .map_err(|error| ParseError::from_file_and_parse_error(file, error))
     }
 }
